api/pyproject.toml
[project]
name = "axis-api"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "argon2-cffi>=25.1.0",
    "cryptography>=46.0.3",
    "email-validator>=2.3.0",
    "fastapi>=0.120.1",
    "passlib[bcrypt]>=1.7.4",
    "psycopg2-binary>=2.9.11",
    "pydantic-settings>=2.11.0",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.20",
    "redis>=7.0.1",
    "sqlalchemy>=2.0.44",
    "uvicorn[standard]>=0.38.0",
]
